refactor(agents): extract the session chat factory into agent/chat
Move the ~200-line imperative Chat construction (transport prep, tool
dispatch, turn completion, conflict handling) out of useAgentChat into
createAgentSessionChat with explicit dependencies. The hook now records
the synced transcript tail through an onTranscriptSynced callback
instead of the factory reaching into the hook's ref, and per-turn client
state is accessed via getTurnClientState. Chat API URL builders, session
conflict error codes, and getRemovedUserMessageText move to small
agent/chat modules shared by both. Pure code motion — no behavior
change.
